+++ package/src/extract/form.rs
@@ -73,7 +73,7 @@ where
             | Ok(val) => Ok(Self(val.0)),
             | Err(rej) => Err(CreateJsonResponse::failure()
                 .status(rej.status())
-                .error(
+                .add_error(
                     JsonResponseError::new()
                         .code(ResponseError::Parse.as_code())
                         .message(rej.body_text()),
@@ -97,7 +97,7 @@ where
                 .body(body),
             | Err(err) => CreateJsonResponse::failure()
                 .status(StatusCode::INTERNAL_SERVER_ERROR)
-                .error(
+                .add_error(
                     JsonResponseError::new()
                         .code(ResponseError::Server.as_code())
                         .message(err.to_string()),