+++ package/src/response/json/mod.rs
@@ -1,12 +1,14 @@
+pub(crate) mod create;
 pub(crate) mod error;
-pub(crate) mod failure;
-pub(crate) mod main;
-pub(crate) mod success;
-
-pub use crate::response::json::main::{CreateJsonResponse, JsonResponse};
-
-pub use crate::response::json::success::JsonSuccessResponseFunctions;
-
-pub use crate::response::json::failure::JsonFailureResponseFunctions;
+pub(crate) mod functions;
+pub(crate) mod response;
 
 pub use crate::response::json::error::{JsonResponseError, ResponseError};
+
+pub use crate::response::json::response::JsonResponse;
+
+pub use crate::response::json::functions::success::JsonSuccessResponseFunctions;
+
+pub use crate::response::json::functions::failure::JsonFailureResponseFunctions;
+
+pub use crate::response::json::create::CreateJsonResponse;